[outputs] refs #47 - Move Format method for thousand commas to util and update packages in modelWallets.
